backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
yfinance
requests
python-multipart
openpyxl
python-dotenv
pytest
